pyproject.toml
[project]
name = "fastapi-migration-skeleton"
version = "0.1.0"
description = "FastAPI project skeleton for ASP.NET Core API migration"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "sqlalchemy>=2.0.30,<3.0.0",
    "alembic>=1.13.0,<2.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "aiosqlite>=0.20.0,<1.0.0",
    "bcrypt>=4.1.0,<5.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "pydantic-settings>=2.4.0,<3.0.0",
    "cloudinary>=1.41.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0,<9.0.0",
]

[project.scripts]
app-seed = "app.db.seed:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
